backend/requirements.txt
# Core Framework
controlflow>=0.12.0
flask>=2.3.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
whitenoise>=6.5.0

# LLM Providers
openai>=1.12.0
google-generativeai>=0.3.0
langchain-google-genai>=1.0.0

# Data & Utilities
requests>=2.31.0
PyYAML>=6.0.0
pydantic>=2.5.0
tenacity>=8.2.0

# Development & Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0